reduce string allocations in environment get/set
