+++ cpp/src/parser/environment.hpp
@@ -2,7 +2,8 @@
 
 #include <memory>
 #include <string_view>
-#include <unordered_map>
+
+#include "utils.hpp"
 
 #include "value.hpp"
 
@@ -12,10 +13,10 @@ using EnvironmentP = std::shared_ptr<Environment>;
 struct Environment : public std::enable_shared_from_this<Environment>
 {
 	EnvironmentP parent{};
-	std::unordered_map<std::string, Value> values;
+	unordered_string_map<Value> values;
 
 	Environment(EnvironmentP parent = {});
 
 	Value get(std::string_view name) const;
-	void set(std::string_view name, Value value);
+	void set(std::string_view name, Value&& value);
 };
